+++ tests/testthat/test-imports.R
@@ -1,0 +1,5 @@
+context("Silly tests for imports.R")
+
+test_that("release_questions", {
+  expect_equal(length(netdiffuseR:::release_questions()), 4L)
+})